requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
gunicorn==23.0.0

uvicorn[standard]
pydantic-settings

fastapi

sqlmodel
pydantic
python-dotenv
openai

requests
psutil
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
aiofiles
python-multipart
